spotify_screen/requirements.txt
Adafruit-Blinka==8.0.2
adafruit-circuitpython-busdevice==5.1.10
adafruit-circuitpython-framebuf==1.4.11
adafruit-circuitpython-ssd1306==2.12.6
adafruit-circuitpython-typing==1.7.1
Adafruit-PlatformDetect==3.24.1
Adafruit-PureIO==1.1.9
certifi==2020.6.20
chardet==4.0.0
click==8.1.3
colorzero==1.1
distro==1.5.0
Flask==2.1.2
gpiozero==1.6.2
idna==2.10
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
olefile==0.46
Pillow==8.1.2
pyftdi==0.54.0
pyserial==3.5
python-apt==2.2.1
pyusb==1.2.1
requests==2.25.1
rpi-ws281x==4.3.4
RPi.GPIO==0.7.0
six==1.16.0
spidev==3.5
ssh-import-id==5.10
sysv-ipc==1.1.0
urllib3==1.26.5
Werkzeug==2.1.2
zipp==3.8.0
